requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
plotly
langgraph
langchain
langchain-groq
python-dotenv
chromadb
sentence-transformers
fpdf2
